requirements.txt
selenium
google-generativeai
Pillow
